+++ lieu/lieux.php
@@ -123,8 +123,10 @@ include("../_header.inc.php");
         <div>
             <div class="table-filters">
                 <form action="" method="get">
-                    <input type="search" name="nom" value="<?= sanitizeForHtml($_SESSION['user_prefs_lieux_nom']) ?>" placeholder="Nom" aria-label="Nom">
-                    <button type="button" class="js-clear-search-field" aria-label="Vider et relancer la recherche" title="Vider et relancer la recherche">&times;</button>
+                    <span class="search-field">
+                        <input type="search" name="nom" value="<?= sanitizeForHtml($_SESSION['user_prefs_lieux_nom']) ?>" placeholder="Nom" aria-label="Nom">
+                        <button type="button" class="js-clear-search-field" aria-label="Vider et relancer la recherche" title="Vider et relancer la recherche"></button>
+                    </span>
                     <select name="categorie" class="js-select2-options-with-style" data-placeholder="Catégorie" style="width:80px">
                          <option value="" placeholder="type"></option>
                         <?php foreach ($glo_categories_lieux as $k => $label) : ?>